+++ libraries/apollo-api/build.gradle.kts
@@ -15,6 +15,12 @@ kotlin {
         api(project(":apollo-annotations"))
       }
     }
+
+    findByName("jsMain")?.apply {
+      dependencies {
+        implementation(libs.okio.nodefilesystem)
+      }
+    }
   }
 }
