+++ libraries/apollo-api/src/concurrentMain/kotlin/com/apollographql/apollo3/api/systemFileSystem.concurrent.kt
@@ -1,6 +1,7 @@
 package com.apollographql.apollo3.api
 
 import okio.FileSystem
+import okio.SYSTEM
 
 internal actual val systemFileSystem: FileSystem
   get() = FileSystem.SYSTEM
