+++ libraries/apollo-api/src/filesystemMain/kotlin/com/apollographql/apollo3/api/DefaultUpload.filesystem.kt
@@ -1,0 +1,22 @@
+@file:JvmMultifileClass
+@file:JvmName("DefaultUploadKt")
+package com.apollographql.apollo3.api
+
+import okio.FileSystem
+import okio.Path
+import okio.buffer
+import okio.use
+import kotlin.jvm.JvmMultifileClass
+import kotlin.jvm.JvmName
+
+fun Path.toUpload(contentType: String, fileSystem: FileSystem = systemFileSystem): Upload {
+  return DefaultUpload.Builder()
+      .content { sink ->
+        fileSystem.openReadOnly(this).use {
+          sink.writeAll(it.source().buffer())
+        }
+      }
+      .contentType(contentType)
+      .contentLength(fileSystem.metadata(this).size ?: -1L)
+      .build()
+}
