--- libraries/apollo-api/src/linuxMain/kotlin/com/apollographql/apollo3/api/systemFileSystem.kt
@@ -1,6 +1,0 @@
-package com.apollographql.apollo3.api
-
-import okio.FileSystem
-
-internal actual val systemFileSystem: FileSystem
-  get() = FileSystem.SYSTEM
