--- libraries/apollo-api/src/wasmJsMain/kotlin/com/apollographql/apollo3/api/-FileSystem.wasmJs.kt
@@ -1,6 +1,0 @@
-package com.apollographql.apollo3.api
-
-import okio.FileSystem
-
-internal actual val systemFileSystem: FileSystem
-  get() = throw IllegalStateException("There is no SYSTEM filesystem on JS")
