+++ libraries/apollo-api/src/wasmJsMain/kotlin/com/apollographql/apollo3/api/systemFileSystem.wasmJs.kt
@@ -3,4 +3,4 @@ package com.apollographql.apollo3.api
 import okio.FileSystem
 
 internal actual val systemFileSystem: FileSystem
-  get() = FileSystem.SYSTEM
+  get() = throw IllegalStateException("There is no SYSTEM filesystem on wasmJs")
